Removed old language handler
